move: emulator functions => emulator folder
