+++ bc4py/contract/emulator/virtualmachine.py
